+++ backend/src/resolvers/Mutation/System/webRule.ts
@@ -1,27 +1,37 @@
 import db from '../../../config/db';
 import { webRule as getWebRule } from '../../Query/System/webRule';
 
-const createWebRule = async (_, { data }) => {
+const createWebRule = async (_, { data }, ctx) => {
+    const webClient = ctx.getWebClient();
+    if (!webClient || !ctx.hasPermission('RULE_SUPERADMIN')) {
+        return new Error('access denied');
+    }
+
     try {
         // Convert webTool in a web_tool_id and delete webTool in the data structure.
         data.web_tool_id = data.webTool.id;
         delete data.webTool;
 
         return db('web_rule').insert(data)
-            .then(id => getWebRule(_, { filter: { id } }))
+            .then(id => getWebRule(_, { filter: { id } }, ctx))
             .catch(e => new Error(e.sqlMessage));
     } catch (e) {
         throw new Error(e.sqlMessage);
     }
 }
 
-const updateWebRule = async (_, { filter, data }) => {
+const updateWebRule = async (_, { filter, data }, ctx) => {
+    const webClient = ctx.getWebClient();
+    if (!webClient || !ctx.hasPermission('RULE_SUPERADMIN')) {
+        return new Error('access denied');
+    }
+
     try {
         // Convert webTool in a web_tool_id and delete webTool in the data structure.
         data.web_tool_id = data.webTool.id;
         delete data.webTool;
 
-        const webRule = await getWebRule(_, { filter });
+        const webRule = await getWebRule(_, { filter }, ctx);
         if (webRule) {
             const { id } = webRule;
             await db('web_rule').where({ id }).update(data)
@@ -32,9 +42,14 @@ const updateWebRule = async (_, { filter, data }) => {
     }
 }
 
-const deleteWebRule = async (_, { filter }) => {
+const deleteWebRule = async (_, { filter }, ctx) => {
+    const webClient = ctx.getWebClient();
+    if (!webClient || !ctx.hasPermission('RULE_SUPERADMIN')) {
+        return new Error('access denied');
+    }
+
     try {
-        const webRule = await getWebRule(_, { filter });
+        const webRule = await getWebRule(_, { filter }, ctx);
         if (webRule) {
             const { id } = webRule;
             await db('web_rule').where({ id }).delete();
